% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_class.R
\docType{class}
\name{anrep-class}
\alias{anrep}
\alias{anrep-class}
\title{Report class}
\description{
Essentially a logger for tables, plots and other objects,
generating a Markdown text that is converted into a multi-page static Web site
by the \code{save} method
}
\section{Methods}{

\describe{
\item{\code{initialize(author = "Anonymous", title = "Analysis", date = base::date(),
  out.file = "report", out.formats = c("html"), incremental.save = F,
  portable.html = T, ...)}}{Construct new instance}
}}

