% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{anrep.anchor.return}
\alias{anrep.anchor.return}
\title{Return Markdown string with a generated anchor (using HTML tag)}
\usage{
anrep.anchor.return(anchor, text)
}
\arguments{
\item{anchor}{Name of anchor}

\item{text}{Title of anchor}
}
\description{
Return Markdown string with a generated anchor (using HTML tag)
}
\examples{
anrep.anchor.return("My_anchor","My anchor title")
}

