% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{anrep.as.printed.return}
\alias{anrep.as.printed.return}
\title{Return Markdown string with the argument printed verbatim}
\usage{
anrep.as.printed.return(x, attrs = "")
}
\arguments{
\item{x}{String or object to print}

\item{attrs}{String with extra Markdown attributes}
}
\description{
Return Markdown string with the argument printed verbatim
}
\examples{
anrep.as.printed.return("**Text**")
}
