% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{anrep.link.verbatim.return}
\alias{anrep.link.verbatim.return}
\title{Return Markdown link, escaping the link title}
\usage{
anrep.link.verbatim.return(url, text = NULL)
}
\arguments{
\item{url}{Link URL}

\item{text}{Link title}
}
\description{
Return Markdown link, escaping the link title
}
\examples{
anrep.link.verbatim.return("my_link","**link title**")
}
