% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anrepr.R
\docType{package}
\name{anrepr-package}
\alias{anrepr}
\alias{anrepr-package}
\title{Analysis Reporting in R}
\description{
Anrepr provides a report generator that acts as a graphical logger
turning the output of your analysis pipeline into a static Web site
}
\details{
The main entry point is the constructor of the \code{\link{anrep}}
class.

\strong{INSTALLATION NOTE:} This package needs the Pandoc Markdown
converter executable installed on your system. You can refer to the documentation of the
\href{https://CRAN.R-project.org/package=pander}{\pkg{pander}} package for
the alternative ways of installing Pandoc.
}

