% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{arg_list_as_str}
\alias{arg_list_as_str}
\title{Return a named list as a string of named function arguments}
\usage{
arg_list_as_str(x, collapse = ",")
}
\arguments{
\item{x}{Named list}

\item{collapse}{Passed to \code{\link[base]{paste}}}
}
\description{
Useful in creating captions where you need to report the parameters
with wich some method was called by the user
}
\examples{
arg_list_as_str(list(a=1,b="bbb"))

}

