% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_util.R
\name{calc.text.size}
\alias{calc.text.size}
\title{Adopted from phyloseq code
Computes text size of axis label based on the number of
labels.
Maybe R strwidth can be used even with ggplot2?}
\usage{
calc.text.size(n, mins = 0.5, maxs = 4, B = 6, D = 100)
}
\description{
Adopted from phyloseq code
Computes text size of axis label based on the number of
labels.
Maybe R strwidth can be used even with ggplot2?
}

