% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_util.R
\name{citation.to.file}
\alias{citation.to.file}
\title{Write citations for a vector of package names into file in BibTex format
TODO can just use bibtex::write.bib
TODO after writing (or before if BibTex allows that type), optionally replace
Manual type with TechReport that Zotero understands in BibTex (converts to Report).
Otherwise Zotero
replaces Manual with Book. The replacement parameter should be a list of
to:from tuples. In the path BibTex -> Zotero -> RIS -> Endnote Web Page
gets converted to Journal Article still. It seems that in certain styles in
EndNote (ACS), the only way too show URL is to set type to Web Page. Otherwise
it is not clear at all that packages are CRAN packages.}
\usage{
citation.to.file(package, file.name, append = F, ...)
}
\description{
Write citations for a vector of package names into file in BibTex format
TODO can just use bibtex::write.bib
TODO after writing (or before if BibTex allows that type), optionally replace
Manual type with TechReport that Zotero understands in BibTex (converts to Report).
Otherwise Zotero
replaces Manual with Book. The replacement parameter should be a list of
to:from tuples. In the path BibTex -> Zotero -> RIS -> Endnote Web Page
gets converted to Journal Article still. It seems that in certain styles in
EndNote (ACS), the only way too show URL is to set type to Web Page. Otherwise
it is not clear at all that packages are CRAN packages.
}

