% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{\%anrep++\%}
\alias{\%anrep++\%}
\title{Infix operator to wrap a report code block after incrementing section index (Header 1.1 ++ Header 1.2)}
\usage{
header_call \%anrep++\% expr
}
\arguments{
\item{header_call}{Call to \code{anrep$header()}}

\item{expr}{Code section that will be reporting under the new subsection}
}
\description{
This is not strictly necessary to use because default \code{anrep$header()}
parameters result in incrementing the section index, and no cleanup handling
is required to end the section, but using this will make the code's relationship
with reporting section more clear, and protect from future implementation changes
if they introduce the internal cleanup code
}
\examples{
## See \code{\link{anrep}} examples
}
