% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{\%anrep//\%}
\alias{\%anrep//\%}
\title{Infix operator to wrap a sub-report code block (Header 1.1 // Subreport Header 1.1.1)}
\usage{
header_call \%anrep//\% expr
}
\arguments{
\item{header_call}{Call to \code{anrep$header()}}

\item{expr}{Code section that be reporting under the new subsection}
}
\description{
Infix operator to wrap a sub-report code block (Header 1.1 // Subreport Header 1.1.1)
}
\examples{
## See \\code{\\link{anrep}} examples
}

