% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{set_default_external_options}
\alias{set_default_external_options}
\title{Reset various global options to tune up a default style of the reports}
\usage{
set_default_external_options()
}
\description{
This overrides various options in the \code{pander} package and changes
the default ggplot2 theme.
}
